pyproject.toml
[project]
name = "evidence-forecast"
version = "0.1.0"
description = "Probabilistic forecast of meta-analysis conclusion flips"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "scikit-learn>=1.5",
    "xgboost>=2.1",
    "pyyaml>=6.0",
    "jsonschema>=4.22",
    "jinja2>=3.1",
    "shap>=0.46",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["evidence_forecast*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
